backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
PyJWT==2.8.0
